Move reuseable helpers in transaction_tests to separate class
